Lazily merge the platform specific code into DbConnectionOptions.cs
